Reimplement MessageService and decrease service coupling

+++ ForumTask.BLL/DTO/TopicDTO.cs
@@ -27,8 +27,8 @@ namespace ForumTask.BLL.DTO
         public long Id { get; set; }
         public string Title { get; set; }
         public DateTime CreateTime { get; set; }
-        public int? CreatorId { get; set; }
+        public long? CreatorId { get; set; }
 
-        public int MessageCount { get; set; }
+        public long MessageCount { get; set; }
     }
 }
